This commit demonstrates an implementation of pulling URI prefixes from ENV vars and not creating any storage for those prefixes.
There are situations where we want to use API Routes within NiceGUI, but have no need for storage. If we create the connections at all we have to manage them after.

See this comment on Issue #4979 to see the demonstration instructions for this branch:

https://github.com/zauberzeug/nicegui/issues/4979#issuecomment-3111515242

+++ nicegui/persistence/redis_persistent_dict.py
@@ -1,6 +1,7 @@
 from .. import background_tasks, core, json, optional_features
 from ..logging import log
 from .persistent_dict import PersistentDict
+from typing import Optional
 
 try:
     import redis as redis_sync
@@ -12,32 +13,56 @@ except ImportError:
 
 class RedisPersistentDict(PersistentDict):
 
-    def __init__(self, *, url: str, id: str, key_prefix: str = 'nicegui:') -> None:  # pylint: disable=redefined-builtin
+    def __init__(self, *, url: str, id: str, key_prefix: str = 'nicegui:',
+                 skip_redis_publish:  Optional[bool] = False) -> None:  # pylint: disable=redefined-builtin
+                # TODO: is Optional required for skip_redis_publish? I think this is only used in one location.
         if not optional_features.has('redis'):
             raise ImportError('Redis is not installed. Please run "pip install nicegui[redis]".')
         self.url = url
-        self.redis_client = redis.from_url(
-            url,
+        self.key = key_prefix + id
+        self.skip_redis_publish = skip_redis_publish
+        self.redis_client = None
+        self.pubsub = None
+        super().__init__(data={}, on_change=self.publish)
+
+    async def _init_redis_client(self):
+        if not self.redis_client:
+            self.redis_client = redis.from_url(
+                self.url,
+                health_check_interval=10,
+                socket_connect_timeout=5,
+                retry_on_timeout=True,
+                socket_keepalive=True,
+            )
+
+    async def _init_pubsub(self):
+        if not self.redis_client:
+            await self._init_redis_client()
+        if not self.pubsub:
+            self.pubsub = self.redis_client.pubsub()
+
+    async def initialize(self) -> None:
+        '''Load initial data from Redis and start listening for changes.'''
+        if self.skip_redis_publish:
+            print(f'⚠️ SKIPPING INIT ON {self.key} because URI PREFIX EXEMPTED') # TODO: remove before pull request commit
+            return
+        async with redis.from_url(
+            self.url,
             health_check_interval=10,
             socket_connect_timeout=5,
             retry_on_timeout=True,
             socket_keepalive=True,
-        )
-        self.pubsub = self.redis_client.pubsub()
-        self.key = key_prefix + id
-        super().__init__(data={}, on_change=self.publish)
-
-    async def initialize(self) -> None:
-        """Load initial data from Redis and start listening for changes."""
-        try:
-            data = await self.redis_client.get(self.key)
-            self.update(json.loads(data) if data else {})
-            self._start_listening()
-        except Exception:
-            log.warning(f'Could not load data from Redis with key {self.key}')
+        ) as redis_client_sync:
+            try:
+                data = await redis_client_sync.get(self.key)
+                self.update(json.loads(data) if data else {})
+                self._start_listening()
+            except Exception:
+                log.warning(f'Could not load data from Redis with key {self.key}')
 
     def initialize_sync(self) -> None:
-        """Load initial data from Redis and start listening for changes in a synchronous context."""
+        '''Load initial data from Redis and start listening for changes in a synchronous context.'''
+        print(f'.... initialize_sync() {self.key}')  # TODO: remove before pull request commit
         with redis_sync.from_url(
             self.url,
             health_check_interval=10,
@@ -50,10 +75,13 @@ class RedisPersistentDict(PersistentDict):
                 self.update(json.loads(data) if data else {})
                 self._start_listening()
             except Exception:
-                log.warning(f'Could not load data from Redis with key {self.key}')
+                log.warning(f'Could not load data from Redis with key {self.key} - {self}')
 
     def _start_listening(self) -> None:
         async def listen():
+            print(f'✅ SUBSCRIBING TO {self.key} - {self}')  # TODO: remove before pull request commit
+            if not self.pubsub:
+                await self._init_pubsub()
             await self.pubsub.subscribe(self.key + 'changes')
             async for message in self.pubsub.listen():
                 if message['type'] == 'message':
@@ -67,26 +95,48 @@ class RedisPersistentDict(PersistentDict):
             core.app.on_startup(listen())
 
     def publish(self) -> None:
-        """Publish the data to Redis and notify other instances."""
+        '''Publish the data to Redis and notify other instances.'''
         async def backup() -> None:
-            pipeline = self.redis_client.pipeline()
-            pipeline.set(self.key, json.dumps(self))
-            pipeline.publish(self.key + 'changes', json.dumps(self))
-            await pipeline.execute()
+            async with redis.from_url(
+                    self.url,
+                    health_check_interval=10,
+                    socket_connect_timeout=5,
+                    retry_on_timeout=True,
+                    socket_keepalive=True,
+            ) as redis_client:
+                pipeline = redis_client.pipeline()
+                pipeline.set(self.key, json.dumps(self))
+                pipeline.publish(self.key + 'changes', json.dumps(self))
+                await pipeline.execute()
+        if self.skip_redis_publish:
+            print(f'⚠️ SKIPPING PUBLISH ON {self.key} because URI PREFIX EXEMPTED') # TODO: remove before pull request commit
+            return
         if core.loop:
             background_tasks.create_lazy(backup(), name=f'redis-{self.key}')
         else:
             core.app.on_startup(backup())
 
     async def close(self) -> None:
-        """Close Redis connection and subscription."""
-        await self.pubsub.unsubscribe()
-        await self.pubsub.close()
-        await self.redis_client.close()
+        '''Close Redis connection and subscription.'''
+        if self.skip_redis_publish:
+            print(f'⚠️ SKIPPING CLOSE ON {self.key} because URI PREFIX EXEMPTED') # TODO: remove before pull request commit
+            return
+        if self.pubsub:
+            await self.pubsub.unsubscribe()
+            await self.pubsub.close()
+        if self.redis_client:
+            await self.redis_client.close()
 
     def clear(self) -> None:
         super().clear()
+        if self.skip_redis_publish:
+            print(f'⚠️ SKIPPING REDIS CLEAR ON {self.key} because URI PREFIX EXEMPTED') # TODO: remove before pull request commit
+            return
+        else: # TODO: remove before pull request commit
+            print(f'✅ CLEARING REDIS KEY {self.key} - {self}') # TODO: remove before pull request commit
         if core.loop:
-            background_tasks.create_lazy(self.redis_client.delete(self.key), name=f'redis-delete-{self.key}')
+            if self.redis_client:
+                background_tasks.create_lazy(self.redis_client.delete(self.key), name=f'redis-delete-{self.key}')
         else:
-            core.app.on_startup(self.redis_client.delete(self.key))
+            if self.redis_client:
+                core.app.on_startup(self.redis_client.delete(self.key))
